+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 + Remove `GenericCaseTree`, add `GenericCase` because now each case in the
   case tree have their own telescope
++ Remove `DeclarationType`, use `is_recursive: bool` instead
 
 # 0.2.5